+++ tests/testthat/test-cycles.R
@@ -1,0 +1,16 @@
+context("fundamental cycles")
+
+test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true") |
+             identical (Sys.getenv ("TRAVIS"), "true"))
+
+test_that("dodgr_fundamental_cycles", {
+              net <- weight_streetnet (hampi)
+              graph <- dodgr_contract_graph (net)$graph
+              expect_error (x <- dodgr_fundamental_cycles (),
+                            "graph must be provided")
+              expect_error (x <- dodgr_fundamental_cycles (graph = "a"),
+                            "graph must be a data.frame object")
+              expect_silent (x <- dodgr_fundamental_cycles (graph))
+              expect_is (x, "list")
+              expect_length (x, 43)
+             })